Added Login, some frontend, backend for tags on Pitems TODO:frontend Pitems

+++ database/migrations/2023_11_16_113734_portfolio_items_tags.php
@@ -1,5 +1,4 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -11,21 +10,24 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('portfolio_items_tags', function (Blueprint $table)
-        {
-            $table->bigIncrements('id');
-//portfolio items
-            $table->bigInteger('portfolio_item_id')->unsigned();
-            $table->foreign('portfolio_item_id')->references('id')
-                    ->on('portfolio_items')
+        // Check if the table already exists
+        if (!Schema::hasTable('portfolio_items_tags')) {
+            Schema::create('portfolio_items_tags', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                
+                //portfolio items
+                $table->bigInteger('portfolio_item_id')->unsigned();
+                $table->foreign('portfolio_item_id')
+                    ->references('id')->on('portfolio_items')
                     ->onDelete('cascade');
-//tags 
-            $table->bigInteger('tag_id')->unsigned();
-            $table->foreign('tag_id')->references('id')
-                    ->on('tags')
+                
+                //tags 
+                $table->bigInteger('tag_id')->unsigned();
+                $table->foreign('tag_id')
+                    ->references('id')->on('tags')
                     ->onDelete('cascade');
-
-        });
+            });
+        }
     }
 
     /**
@@ -33,6 +35,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        //
+        Schema::dropIfExists('portfolio_items_tags');
     }
 };
